Added comments, design stuff, reorganized folder structure
